Greedy Fix
I am relatively sure that I fixed the greedy algorithm. I have not done extensive testing, but from what I have done, everything seems to be working right. The structure we had was good, but there were some extra artifacts (due to the fact that we changed how we were doing it several times midway) that I cleaned up. I am not entirely sure what I did that made the difference. It turns out we forgot to add a fail condition (so that may have been it). Another theory is that because we had several different variables floating around, one of them was misused.

I made two small changes to the GUI file. First, I changed the minimum size to be smaller to fit my screen better (we can change it back if necessary). Second, I added some exception handling to make sure the error is printed before a crash.