Fixed the incorrect connections
It turns out that it was a simple problem to fix. The graph is a digraph (so edges are not bidirectional) and we were accidentally adding i to j when we checked for j to i.
I also made the fancy algorithm the default one on the drop down for our ease of use.