Fixed back tracking, small construction bug
I did primarily two things. First, I made the backtracking work. It turns out, that I didn't convert it completely from the connection greedy, so that should be fixed now. There are some scenarios that normal greedy may fail on (for example, Hard deterministic seed 22), but even so, the secure greedy usage will be guaranteed to get some path. This is essential for the fancy to work correctly.

There was also a small bug in the path construction for the fancy algorithm. I had forgotten to use % in one place, and thus it had the chance to crash. That has now been resolved.